# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Whirlpool Inspection Tool API",
    "description": "Whirlpool Inspection Tool API APIs developed by Scopt Analytics.",
    "contact": {
      "name": "Scopt Analytics",
      "url": "https://scoptanalytics.com/",
      "email": "support@scoptanalytics.com"
    },
    "license": {
      "name": "Proprietary - Internal Use Only",
      "url": "https://scoptanalytics.com/terms"
    },
    "version": "1.0.3"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Complete login using SSO exchange token from Okta callback",
        "operationId": "login_token_auth_login_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot Password",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/devices/resolve": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Keep selected devices and deregister the rest (single-login mode)",
        "operationId": "resolve_auth_devices_auth_devices_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDevicesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveDevicesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/devices/{device_uuid}/deregister": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Deregister a device and revoke its sessions",
        "operationId": "deregister_auth_device_auth_devices__device_uuid__deregister_post",
        "parameters": [
          {
            "name": "device_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeregisterDeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActiveDeviceResponse": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "imei": {
            "type": "string",
            "title": "Imei"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type"
          },
          "device_fingerprint": {
            "type": "string",
            "title": "Device Fingerprint"
          },
          "display_label": {
            "type": "string",
            "title": "Display Label"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current",
            "default": false
          },
          "has_active_session": {
            "type": "boolean",
            "title": "Has Active Session",
            "default": false
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "imei",
          "device_type",
          "device_fingerprint",
          "display_label"
        ],
        "title": "ActiveDeviceResponse"
      },
...
      "DeregisterDeviceResponse": {
        "properties": {
          "device_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Device Uuid"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Device deregistered and sessions revoked"
          }
        },
        "type": "object",
        "required": [
          "device_uuid"
        ],
        "title": "DeregisterDeviceResponse"
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest"
      },
      "ForgotPasswordResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ForgotPasswordResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginDeviceInfo": {
        "properties": {
          "imei": {
            "type": "string",
            "maxLength": 48,
            "minLength": 3,
            "title": "Imei"
          },
          "device_type": {
            "type": "string",
            "pattern": "^(desktop|mobile)$",
            "title": "Device Type"
          },
          "device_fingerprint": {
            "type": "string",
            "maxLength": 255,
            "minLength": 10,
            "title": "Device Fingerprint"
          },
          "device_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Device Info",
            "description": "JSON object with device / browser info from PWA (userAgent, platform, etc.)"
          },
          "current_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Lat"
          },
          "current_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Lng"
          }
        },
        "type": "object",
        "required": [
          "imei",
          "device_type",
          "device_fingerprint",
          "device_info"
        ],
        "title": "LoginDeviceInfo"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoginDeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "designation": {
            "type": "string",
            "title": "Designation"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "device_uuid": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Uuid"
          },
          "allowed_warehouses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Warehouses",
            "description": "Warehouse codes the user may access; null for superadmin."
          },
          "allow_multi_login": {
            "type": "boolean",
            "title": "Allow Multi Login",
            "default": true
          },
          "requires_device_selection": {
            "type": "boolean",
            "title": "Requires Device Selection",
            "default": false
          },
          "active_devices": {
            "items": {
              "$ref": "#/components/schemas/ActiveDeviceResponse"
            },
            "type": "array",
            "title": "Active Devices"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uuid",
          "name",
          "email",
          "role",
          "designation",
          "is_active",
          "access_token"
        ],
        "title": "LoginResponse"
      },
      "LoginTokenRequest": {
        "properties": {
          "access_token": {
            "type": "string",
            "minLength": 10,
            "title": "Access Token",
            "description": "Short-lived SSO exchange token from the Okta callback redirect"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LoginDeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "LoginTokenRequest"
      },
...
      "ResolveDevicesRequest": {
        "properties": {
          "keep_device_uuids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Keep Device Uuids",
            "description": "Device UUIDs to keep active; all other active devices are deregistered."
          }
        },
        "type": "object",
        "required": [
          "keep_device_uuids"
        ],
        "title": "ResolveDevicesRequest"
      },
      "ResolveDevicesResponse": {
        "properties": {
          "kept_device_uuids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Kept Device Uuids"
          },
          "deregistered_device_uuids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Deregistered Device Uuids"
          }
        },
        "type": "object",
        "required": [
          "kept_device_uuids",
          "deregistered_device_uuids"
        ],
        "title": "ResolveDevicesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}